+++ packages/rapier/src/components/colliders/CylinderCollider.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+import Collider from './BaseCollider.vue'
+import type { ColliderProps } from '../../types'
+
+const props = defineProps<Partial<Omit<ColliderProps, 'shape'>>>()
+</script>
+
+<template>
+  <Collider shape="cylinder" v-bind="props" />
+</template>